# perf(cli): use jemalloc to get faster research time with musl