--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -33,3 +33,6 @@ serde_regex = "1.1"
 serde_yaml = "0.8"
 tokio = { version = "=1.20", features = ["fs", "io-util", "macros", "rt-multi-thread"] }
 toml = "0.5"
+
+[target.'cfg(not(target_env = "msvc"))'.dependencies]
+tikv-jemallocator = "0.5"