Running compiled `wakuchin_cli` for `x86_64-unknown-linux-musl` with `-i 2000000 -r WKNCWKNC -t 2` on my machine:

```
Benchmark 1: ./wakuchin-jemalloc -i 2000000 -r WKNCWKNC -t 2
  Time (mean ± σ):     463.5 ms ±   6.6 ms    [User: 3553.5 ms, System: 43.9 ms]
  Range (min … max):   454.0 ms … 472.9 ms    10 runs
 
Benchmark 2: ./wakuchin-nojemalloc -i 2000000 -r WKNCWKNC -t 2
  Time (mean ± σ):      7.253 s ±  0.147 s    [User: 58.133 s, System: 35.767 s]
  Range (min … max):    6.954 s …  7.431 s    10 runs
```